Transparent wall activation created
Press X to make the walls completely transparent for 5 seconds. Once 5 seconds pass, the walls reset. This can be done 5 times. According to Unity documentation, I used a legacy shader for this process, but the new way is to create a transparent material using the Standard Shader (https://docs.unity3d.com/Manual/shader-StandardShader.html). If we think this might make a difference, we can re-implement using the transparent material that I created in the Assets/Materials folder.